Guard against silent (path,method)/operationId collisions; dedupe two copies
- gen-openapi.mjs: fail loud on a duplicate (path, method) across two
  GROUPS, and on a duplicate operationId — both previously would have
  silently overwritten the earlier entry or shipped an invalid document,
  matching this file's existing fail-loud convention for unmapped types
  and auth levels. 0 collisions today; this only guards the future.
- New scripts/renderFilterSection.mjs: the "Filtering jobs" section
  (modifiers, facet/extras tables, recipes) was hand-built twice,
  independently, in gen-api-docs.mjs and gen-openapi.mjs — the exact
  kind of drift this migration's single-source-of-truth goal exists to
  eliminate. Both generators now call the one shared function.
  docs/API.md is byte-identical; the OpenAPI document gains one
  previously-markdown-only sentence ("Every facet below supports
  repeat-OR...") it was missing.
- New scripts/deprecatedFixture.mjs: the synthetic fixture exercising
  the `deprecated` field (no shipped endpoint qualifies yet) was
  duplicated near-verbatim across both smoke tests; now built once.
